Правки отрисовки карточек и фильтра
