+++ src/render-filter.js
@@ -1,15 +1,18 @@
-import addFilterClickHandler from './add-filter-click-handler.js';
+import renderElement from './utils.js';
 
 // Рендеринг одного пункта фильтра
-const renderFilter = (filterName, taskAmount, isChecked = false) => {
+const renderFilter = (filterName, taskAmount, isChecked = false, renderTaskList, taskContainer, switchChecked) => {
   const string = `<input type="radio" id="filter__${filterName}" class="filter__input visually-hidden" name="filter" ${isChecked && taskAmount ? ` checked` : ``} ${taskAmount ? `` : ` disabled`}><label for="filter__${filterName}" class="filter__label">${filterName}<span class="filter__${filterName}-count"> ${taskAmount}</span></label>`;
 
-  const template = document.createElement(`template`);
-  template.innerHTML = string;
-
-  const element = template.content.children; // HTML-коллекция
-  addFilterClickHandler(element[0], taskAmount);
-  // console.log(element);
+  const element = renderElement(string);
+  const input = element.querySelector(`input`);
+  
+  // Вешаем обработчик
+  input.addEventListener(`click`, (evt) => {
+    evt.preventDefault();
+    switchChecked(evt.target);
+    renderTaskList(taskAmount, taskContainer);
+  });
 
   return element;
 };
